Export RandomSingularBatch through batch_test_util.go
